feat(): 1) MinioController -> ResourceController (change); 2) create ResourceRequestDTO record class;

+++ src/main/java/io/github/xcvqqz/cloud_file_storage/dto/response/ResourceResponseDTO.java
@@ -4,7 +4,7 @@ public record ResourceResponseDTO(
 
         String path,
         String name,
-        byte size,
+        long size,
         String type
 
 )
